Move sqlite plugins etc into utils/sqlite
